tests/wpt/tests/css/css-break/animation/break-no-interpolation.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/w3c/csswg-drafts/issues/4441">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<body>
<script>
test_no_interpolation({
  property: 'break-after',
  from: 'initial',
  to: 'avoid'
});

test_no_interpolation({
  property: 'break-before',
  from: 'initial',
  to: 'avoid'
});

test_no_interpolation({
  property: 'break-inside',
  from: 'initial',
  to: 'avoid'
});
</script>
